requirements.txt
altair>=5.2,<6.5
numpy>=1.26,<3.0
onnxruntime>=1.17,<2.0
pandas>=2.1,<3.0
Pillow>=10.0,<13.0
requests>=2.31,<3.0
streamlit>=1.36,<2.0
streamlit-drawable-canvas-bogaczm>=0.9.4,<1.0
